Rename files and update spies to mocks
