+++ src/cli/crd/create/createCRDscaffold.test.ts
